Remove logic from components
